feat: Add ShapeUpload as an embedded schema for shape selection
