+++ lib/arrow/shuttle/shape.ex
@@ -1,5 +1,5 @@
 defmodule Arrow.Shuttle.Shape do
-  @moduledoc "schema for shuttle shapes"
+  @moduledoc "schema for shuttle shapes for the db"
   use Ecto.Schema
   import Ecto.Changeset
 
@@ -7,14 +7,12 @@ defmodule Arrow.Shuttle.Shape do
   @type t :: %__MODULE__{
           id: id,
           name: String.t(),
-          coordinates: list(String.t()),
           inserted_at: DateTime.t(),
           updated_at: DateTime.t()
         }
 
   schema "shapes" do
     field :name, :string
-    field :coordinates, {:array, :string}
 
     timestamps(type: :utc_datetime)
   end
@@ -22,7 +20,7 @@ defmodule Arrow.Shuttle.Shape do
   @doc false
   def changeset(shape, attrs) do
     shape
-    |> cast(attrs, [:name, :coordinates])
+    |> cast(attrs, [:name])
     |> validate_required([:name])
     |> unique_constraint(:name)
   end
