+++ lib/arrow_web/controllers/shape_html/show.html.heex
@@ -11,12 +11,5 @@
 <.list>
   <:item title="Name"><%= @shape.name %></:item>
 </.list>
-<.list>
-    <:item title="Coordinates">    
-    <%= for coord <- @shape.coordinates do %>
-       <div> <%= coord %> </div>
-    <% end %> 
-    </:item>
-</.list>
 
 <.back navigate={~p"/shapes"}>Back to shapes</.back>
